Refactor to separate container/component
Also noticed that Delete aria label was not localized.
